pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "navil"
version = "0.2.1"
description = "Security gateway for MCP servers"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [
    {name = "Pantheon Lab Pte Ltd"},
]
keywords = [
    "mcp",
    "security",
    "supply-chain",
    "ai-agents",
    "credentials",
    "policy-engine",
    "anomaly-detection",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: System :: Monitoring",
    "Typing :: Typed",
]

dependencies = [
    "PyYAML>=6.0",
    "PyJWT>=2.8.0",
    "cryptography>=41.0",
    "orjson>=3.9",
    "httpx>=0.27",
    "redis>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2.2",
    "pytest-cov>=4.0",
    "ruff==0.15.6",
    "mypy>=1.8",
    "types-PyYAML>=6.0",
]
ml = [
    "scikit-learn>=1.3",
    "numpy>=1.24",
    "joblib>=1.3",
]
llm = [
    "anthropic>=0.25",
    "openai>=1.12",
    "google-generativeai>=0.5",
]
cloud = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "stripe>=8.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "redis>=5.0",
    "apscheduler>=3.10",
    "resend>=0.7",
]
static-analysis = [
    "tree-sitter>=0.23",
    "tree-sitter-python>=0.23",
    "tree-sitter-javascript>=0.23",
    "tree-sitter-typescript>=0.23",
]
all = [
    "navil[ml]",
    "navil[llm]",
    "navil[cloud]",
    "navil[static-analysis]",
]

[project.scripts]
navil = "navil.cli:main"

[project.urls]
Homepage = "https://github.com/ivanlkf/navil"
Repository = "https://github.com/ivanlkf/navil"
Documentation = "https://github.com/ivanlkf/navil#readme"
Issues = "https://github.com/ivanlkf/navil/issues"
Changelog = "https://github.com/ivanlkf/navil/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
include = ["navil*"]

[tool.setuptools.package-data]
navil = [
    "py.typed",
    "sample_configs/*.json",
    "sample_policies/*.yaml",
    "data/*.yaml",
    "data/*.json",
    "lua/*.lua",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.ruff.lint.per-file-ignores]
"navil/ml/*.py" = ["N806"]   # X, X_scaled are conventional in ML code
"tests/test_ml_*.py" = ["E402", "N806"]  # imports after importorskip; ML X convention
"navil/cloud/email.py" = ["E501"]  # HTML email templates have long lines
"tests/test_adversarial.py" = ["B017", "E402", "E501"]  # security tests: catch any exception; importorskip; long docstrings
"tests/test_routes_validation.py" = ["E402"]  # imports after importorskip

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true

# Optional dependencies — stubs not always available in CI
[[tool.mypy.overrides]]
module = [
    "fastapi.*",
    "uvicorn.*",
    "starlette.*",
    "pydantic.*",
    "httpx.*",
    "anthropic.*",
    "openai.*",
    "google.*",
    "sklearn.*",
    "numpy.*",
    "joblib.*",
    "sqlalchemy.*",
    "alembic.*",
    "stripe.*",
    "redis.*",
    "apscheduler.*",
    "resend.*",
    "psycopg2.*",
    "tree_sitter.*",
    "tree_sitter_python.*",
    "tree_sitter_javascript.*",
    "tree_sitter_typescript.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=navil --cov-report=term-missing --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["navil"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "navil/cloud/*",   # tested via cloud integration tests
    "navil/ml/*",      # tested in separate test-ml CI job
    "navil/cli.py",    # CLI glue code — tested via integration
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
